# RedBlackTree.add() optimization: prevents tree re-creation, ...