--- a/src/main/java/javaslang/collection/RedBlackTree.java
+++ b/src/main/java/javaslang/collection/RedBlackTree.java
@@ -71,67 +71,17 @@ public interface RedBlackTree<T> extends BinarySearchTree<T> {
                     final TreeNode<T> node = (TreeNode<T>) tree;
                     final int comparison = node.comparator.compare(node.value, value);
                     if (comparison > 0) {
-                        return lbalance(node.color, insert(node.left), node.value, node.right, node.comparator);
+                        final RedBlackTree<T> currentLeft = node.left;
+                        final RedBlackTree<T> newLeft = insert(currentLeft);
+                        return (newLeft == currentLeft) ? node : TreeNode.lbalance(node.color, newLeft, node.value, node.right, node.comparator);
                     } else if (comparison < 0) {
-                        return rbalance(node.color, node.left, node.value, insert(node.right), node.comparator);
+                        final RedBlackTree<T> currentRight = node.right;
+                        final RedBlackTree<T> newRight = insert(currentRight);
+                        return (newRight == currentRight) ? node : TreeNode.rbalance(node.color, node.left, node.value, newRight, node.comparator);
                     } else {
                         return node;
                     }
                 }
-            }
-
-            private TreeNode<T> lbalance(Color color, RedBlackTree<T> left, T value, RedBlackTree<T> right, Comparator<? super T> comparator) {
-                if (color == BLACK) {
-                    if (!left.isEmpty()) {
-                        final TreeNode<T> ln = (TreeNode<T>) left;
-                        if (ln.color == RED) {
-                            if (!ln.left.isEmpty()) {
-                                final TreeNode<T> lln = (TreeNode<T>) ln.left;
-                                if (lln.color == RED) {
-                                    final TreeNode<T> newLeft = new TreeNode<>(BLACK, lln.left, lln.value, lln.right, comparator);
-                                    final TreeNode<T> newRight = new TreeNode<>(BLACK, ln.right, value, right, comparator);
-                                    return new TreeNode<>(RED, newLeft, ln.value, newRight, comparator);
-                                }
-                            }
-                            if (!ln.right.isEmpty()) {
-                                final TreeNode<T> lrn = (TreeNode<T>) ln.right;
-                                if (lrn.color == RED) {
-                                    final TreeNode<T> newLeft = new TreeNode<>(BLACK, ln.left, ln.value, lrn.left, comparator);
-                                    final TreeNode<T> newRight = new TreeNode<>(BLACK, lrn.right, value, right, comparator);
-                                    return new TreeNode<>(RED, newLeft, lrn.value, newRight, comparator);
-                                }
-                            }
-                        }
-                    }
-                }
-                return new TreeNode<>(color, left, value, right, comparator);
-            }
-
-            private TreeNode<T> rbalance(Color color, RedBlackTree<T> left, T value, RedBlackTree<T> right, Comparator<? super T> comparator) {
-                if (color == BLACK) {
-                    if (!right.isEmpty()) {
-                        final TreeNode<T> rn = (TreeNode<T>) right;
-                        if (rn.color == RED) {
-                            if (!rn.left.isEmpty()) {
-                                final TreeNode<T> rln = (TreeNode<T>) rn.left;
-                                if (rln.color == RED) {
-                                    final TreeNode<T> newLeft = new TreeNode<>(BLACK, left, value, rln.left, comparator);
-                                    final TreeNode<T> newRight = new TreeNode<>(BLACK, rln.right, rn.value, rn.right, comparator);
-                                    return new TreeNode<>(RED, newLeft, rln.value, newRight, comparator);
-                                }
-                            }
-                            if (!rn.right.isEmpty()) {
-                                final TreeNode<T> rrn = (TreeNode<T>) rn.right;
-                                if (rrn.color == RED) {
-                                    final TreeNode<T> newLeft = new TreeNode<>(BLACK, left, value, rn.left, comparator);
-                                    final TreeNode<T> newRight = new TreeNode<>(BLACK, rrn.left, rrn.value, rrn.right, comparator);
-                                    return new TreeNode<>(RED, newLeft, rn.value, newRight, comparator);
-                                }
-                            }
-                        }
-                    }
-                }
-                return new TreeNode<>(color, left, value, right, comparator);
             }
         }
 
@@ -272,6 +222,60 @@ public interface RedBlackTree<T> extends BinarySearchTree<T> {
 
         private boolean isLeaf() {
             return left.isEmpty() && right.isEmpty();
+        }
+
+        private static <T> TreeNode<T> lbalance(Color color, RedBlackTree<T> left, T value, RedBlackTree<T> right, Comparator<? super T> comparator) {
+            if (color == BLACK) {
+                if (!left.isEmpty()) {
+                    final TreeNode<T> ln = (TreeNode<T>) left;
+                    if (ln.color == RED) {
+                        if (!ln.left.isEmpty()) {
+                            final TreeNode<T> lln = (TreeNode<T>) ln.left;
+                            if (lln.color == RED) {
+                                final TreeNode<T> newLeft = new TreeNode<>(BLACK, lln.left, lln.value, lln.right, comparator);
+                                final TreeNode<T> newRight = new TreeNode<>(BLACK, ln.right, value, right, comparator);
+                                return new TreeNode<>(RED, newLeft, ln.value, newRight, comparator);
+                            }
+                        }
+                        if (!ln.right.isEmpty()) {
+                            final TreeNode<T> lrn = (TreeNode<T>) ln.right;
+                            if (lrn.color == RED) {
+                                final TreeNode<T> newLeft = new TreeNode<>(BLACK, ln.left, ln.value, lrn.left, comparator);
+                                final TreeNode<T> newRight = new TreeNode<>(BLACK, lrn.right, value, right, comparator);
+                                return new TreeNode<>(RED, newLeft, lrn.value, newRight, comparator);
+                            }
+                        }
+                    }
+                }
+            }
+            return new TreeNode<>(color, left, value, right, comparator);
+        }
+
+        private static <T> TreeNode<T> rbalance(Color color, RedBlackTree<T> left, T value, RedBlackTree<T> right, Comparator<? super T> comparator) {
+            if (color == BLACK) {
+                if (!right.isEmpty()) {
+                    final TreeNode<T> rn = (TreeNode<T>) right;
+                    if (rn.color == RED) {
+                        if (!rn.left.isEmpty()) {
+                            final TreeNode<T> rln = (TreeNode<T>) rn.left;
+                            if (rln.color == RED) {
+                                final TreeNode<T> newLeft = new TreeNode<>(BLACK, left, value, rln.left, comparator);
+                                final TreeNode<T> newRight = new TreeNode<>(BLACK, rln.right, rn.value, rn.right, comparator);
+                                return new TreeNode<>(RED, newLeft, rln.value, newRight, comparator);
+                            }
+                        }
+                        if (!rn.right.isEmpty()) {
+                            final TreeNode<T> rrn = (TreeNode<T>) rn.right;
+                            if (rrn.color == RED) {
+                                final TreeNode<T> newLeft = new TreeNode<>(BLACK, left, value, rn.left, comparator);
+                                final TreeNode<T> newRight = new TreeNode<>(BLACK, rrn.left, rrn.value, rrn.right, comparator);
+                                return new TreeNode<>(RED, newLeft, rn.value, newRight, comparator);
+                            }
+                        }
+                    }
+                }
+            }
+            return new TreeNode<>(color, left, value, right, comparator);
         }
     }