if value is already present.

**Simple _Test case_:**

```
public class Test {

    public static void main(String[] args) {

        Random random = Checkable.RNG.get();
        RedBlackTree<Integer> tree = RedBlackTree.empty();

        long start = System.currentTimeMillis();

        for (int i = 0; i < 1_000_000; i++) {
            tree = tree.add(random.nextInt());
        }

        System.out.println("Took " + (System.currentTimeMillis() - start) + " ms");
    }
}
```

Setup:

```
MacBook Pro (15-inch, Early 2011)
Processor 2.2 GHz Intel Core i7
Memory 16 GB 1333 MHz DDR3

Java(TM) SE Runtime Environment (build 1.8.0_51-b16)
Java HotSpot(TM) 64-Bit Server VM (build 25.51-b03, mixed mode)
```

## Before optimization

**Test case:**

```
Took 1964 ms
```

**Property checks:**

```
tree contains no duplicate values: OK, passed 1000 tests in 218 ms.
n := size(tree) => depth(node) <= 2 * floor(log2(n + 1)), for all nodes of tree: OK, passed 1000 tests in 111 ms.
No red node has a red child: OK, passed 1000 tests in 8 ms.
Every path from the root to an empty node contains the same number of black nodes: OK, passed 1000 tests in 108 ms.
```

## After optimization

**Test case:**

```
Took 1573 ms
```

**Property checks:**

```
tree contains no duplicate values: OK, passed 1000 tests in 152 ms.
n := size(tree) => depth(node) <= 2 * floor(log2(n + 1)), for all nodes of tree: OK, passed 1000 tests in 85 ms.
No red node has a red child: OK, passed 1000 tests in 8 ms.
Every path from the root to an empty node contains the same number of black nodes: OK, passed 1000 tests in 74 ms.
```